# go-middleware/cmd/docs/swagger.yaml
basePath: /
definitions:
  model.User:
    properties:
      created_at:
        type: string
      dob:
        type: string
      email:
        description: |-
          FirstName string    `json:"first_name"`
          LastName  string    `json:"last_name"`
        type: string
      id:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
  pkg.ErrorResponse:
    properties:
      errors:
        items:
          type: string
        type: array
      message:
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: golong kominfo 006 api documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: GO DTS USER API DUCUMENTATION
  version: "2.0"
paths:
  /users:
    get:
      consumes:
      - application/json
      description: will fetch 3rd party server to get users data
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
      summary: Show users list
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: will delete user with given id from param
      parameters:
      - description: bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
      summary: Delete user by selected id
      tags:
      - users
    get:
      consumes:
      - application/json
      description: will fetch 3rd party server to get users data to get detail user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
      summary: Show users detail
      tags:
      - users
schemes:
- http
swagger: "2.0"